Part IV/app/requirements.txt
pillow #image processing lib
boto3 #python sdk
pydantic